Build: Clean up development files in production and install self as dep
Distinguish between two use cases:
* Development - i.e., clone the repo, run npm install and then test and debug
* Production - install this package as a dependency of another package

In the development use case, we leave the source files, test scripts, and
everything else lying around after npm install completes.

In the production case, we remove absolutely everything that is not needed for
running the addon - tests, source files, headers, and build scripts.

The reason this package must install itself as if it were a dependency of this
repo is to run the test suites on the built-for-production package. This is
done on the CI server, where the first step is npm pack followed by
npm install *tgz. This will cause a built-for-production package to be
installed under node_modules/. The rest of the dependencies are then installed
with npm install --ignore-scripts.

Fixes gh-88
Fixes gh-89

Signed-off-by: Gabriel Schulhof <gabriel.schulhof@intel.com>